--- Content.Server/VentCraw/Components/BeingVentCrawComponent.cs
@@ -1,8 +1,0 @@
-namespace Content.Server.VentCraw.Components;
-
-[RegisterComponent]
-public sealed class BeingVentCrawComponent : Component
-{
-    [ViewVariables]
-    public EntityUid Holder;
-}
