--- Content.Server/VentCraw/Components/VentCrawBendComponent.cs
@@ -1,7 +1,0 @@
-namespace Content.Server.VentCraw.Components;
-
-[RegisterComponent]
-[Access(typeof(VentCrawTubeSystem))]
-public sealed class VentCrawBendComponent : Component
-{
-}
