--- Content.Server/VentCraw/Components/VentCrawEntryComponent.cs
@@ -1,9 +1,0 @@
-namespace Content.Server.VentCraw.Components
-{
-    [RegisterComponent]
-    [Access(typeof(VentCrawTubeSystem))]
-    public sealed class VentCrawEntryComponent : Component
-    {
-        public const string HolderPrototypeId = "VentCrawHolder";
-    }
-}
