--- Content.Server/VentCraw/Components/VentCrawTransitComponent.cs
@@ -1,9 +1,0 @@
-namespace Content.Server.VentCraw.Components
-{
-    [RegisterComponent]
-    [Access(typeof(VentCrawTubeSystem))]
-    [Virtual]
-    public class VentCrawTransitComponent : Component
-    {
-    }
-}
