+++ Content.Shared/VentCraw/Components/BeingVentCrawComponent.cs
@@ -1,0 +1,8 @@
+namespace Content.Shared.VentCraw.Components;
+
+[RegisterComponent]
+public sealed partial class BeingVentCrawComponent : Component
+{
+    [ViewVariables]
+    public EntityUid Holder;
+}
