+++ Content.Shared/VentCraw/Components/VentCrawBendComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Shared.VentCraw.Components;
+
+[RegisterComponent]
+public sealed partial class VentCrawBendComponent : Component
+{
+}
