+++ Content.Shared/VentCraw/Components/VentCrawEntryComponent.cs
@@ -1,0 +1,7 @@
+namespace Content.Shared.VentCraw.Components;
+
+[RegisterComponent]
+public sealed partial class VentCrawEntryComponent : Component
+{
+    public const string HolderPrototypeId = "VentCrawHolder";
+}
