+++ Content.Shared/VentCraw/Components/VentCrawJunctionComponent.cs
@@ -1,0 +1,10 @@
+namespace Content.Shared.VentCraw.Components;
+
+[RegisterComponent, Virtual]
+public partial class VentCrawJunctionComponent : Component
+{
+    /// <summary>
+    ///     The angles to connect to.
+    /// </summary>
+    [DataField("degrees")] public List<Angle> Degrees = new();
+}
