+++ Content.Shared/VentCraw/Components/VentCrawTransitComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Shared.VentCraw.Components;
+
+[RegisterComponent, Virtual]
+public partial class VentCrawTransitComponent : Component
+{
+}
